extracting some classes from multidim_array
